% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrmap.R
\name{chrmap}
\alias{chrmap}
\title{Map location}
\usage{
chrmap(data, obj = TRUE, exclude = "default", probe = -15000,
  verbose = 1, include = FALSE, ingenes = FALSE)
}
\arguments{
\item{data}{MAList or RGList}

\item{obj}{ligical, create \code{chrom} and \code{maploc} objects}

\item{exclude}{exclude probes: \code{"default"} exclude all probes besides regular probes.
\code{"control"} - exclude only control probes = 0 and 1.
\code{"probe"} - exclude your probes defined in \code{probe} argument.
\code{"none"} - no excluding}

\item{probe}{numeric, which probe you need exclude}

\item{verbose}{logical, system messages}

\item{include}{logical, if TRUE drop any probes described in non-chromosome (1:24,X,Y) regions}

\item{ingenes}{logical, if TRUE create \code{maploc} and \code{chrom} column in result data}
}
\description{
Crosome and map
}

