% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeg.R
\name{commonSeg}
\alias{commonSeg}
\title{Calculate common CNA statistic}
\usage{
commonSeg(data, data2, biomart = FALSE)
}
\arguments{
\item{data}{CBS data, found CNA from \code{thrCBS}}

\item{data2}{DNAcopy data}

\item{biomart}{logical, if TRUE \code{\link{biomaRt}} will be used for description}
}
\description{
Calculate common CNA statistic
}

