% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCGHcall.R
\name{convertCGHcall}
\alias{convertCGHcall}
\title{Convert MAlist to CGHcall}
\usage{
convertCGHcall(x, XY = T, numChr = T)
}
\arguments{
\item{x}{MAList}

\item{XY}{logical, if TRUE exclude XY}

\item{numChr}{number to chr}
}
\description{
Convert MAlist to CGHcall
}

