% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLRSpread.R
\name{DLRSpread}
\alias{DLRSpread}
\title{Derivative log ratio spread}
\usage{
DLRSpread(data)
}
\arguments{
\item{data}{RGList or MAList}
}
\value{
Excellent < 0.2, good < 0.3, bad > 0.3
}
\description{
Derivative log ratio spread (DLRS) 
is a measurement of point-to-point consistency or noisiness in log ratio data
}

