% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutliners.R
\name{findOutliners}
\alias{findOutliners}
\title{Find outliners in arrays}
\usage{
findOutliners(data, smp = 1, flags = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{RGList, EListRaw}

\item{smp}{numeric, number of sample}

\item{flags}{logical, if \code{TRUE} find outliners by flags else by ratio of mean and median}

\item{verbose}{logical, if \code{TRUE} create plot of delta (only if delta is \code{TRUE})}
}
\description{
Exclude outliners by 2 ways, either by FE outliner flags or find spost with unequal 
ratio of mean and median pixel intensity.
}

