% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixArrays.R
\name{fixArrays}
\alias{fixArrays}
\title{Slope correction intra-array}
\usage{
fixArrays(data)
}
\arguments{
\item{data}{RGList, EListRaw}
}
\description{
Correction slope of arrays by Negatives probe and based on Affine transformation. 
Optimal span for 2DLoess normalization by used generalized cross-validation. See 
\code{gcv} in \link{\code{fANCOVA}}.
}

