% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggenomeplot.R
\name{ggenomeplot}
\alias{ggenomeplot}
\title{Genome plot}
\usage{
ggenomeplot(data, smp = NULL, chr = NULL, span = 0.1, ...)
}
\arguments{
\item{data}{MAList or DNAcopy class data}

\item{smp}{numeric, number of sample. if NULL all sample will be saved in current directory}

\item{chr}{1:23, X, Y. If NULL karyotype will be made}

\item{span}{numeric, the parameter α which controls the degree of smoothing.}

\item{...}{params for \code{\link{qplot}} only if param \code{chr} was applied}
}
\description{
Genome plot
}

