% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gplotDG.R
\name{gplotDendro}
\alias{gplotDendro}
\title{Plot dendrogramm for Agilent microarray data}
\usage{
gplotDendro(data, methodDist = "euclidean", methodHclust = "complete",
  rotate = TRUE, title = NULL, ...)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{methodHclust}{method for hierarchical cluster analysis. See \code{\link{hclust}}. Default: {"complete"}.}

\item{rotate}{if TRUE, rotates plot by 90 degrees. Default: {TRUE}.}

\item{title}{title}

\item{...}{see \code{\link{ggdendrogram}}}
}
\description{
Clustering tree by Hierarchical clustering algorithm.
}

