% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotDensities.R
\name{gplotDensities}
\alias{gplotDensities}
\title{Plot Arrays Densities
Plot the density of values for multiple arrays on the same plot.}
\usage{
gplotDensities(data, type = c("density", "violin", "box"))
}
\arguments{
\item{data}{ElistRaw, EList, MAList or RGList}

\item{type}{Type of plot: \code{density}, \code{violin}, \code{box}}
}
\description{
Plot Arrays Densities
Plot the density of values for multiple arrays on the same plot.
}

