% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotKM.R
\name{gplotKMeans}
\alias{gplotKMeans}
\title{Plot clustering MDS for Agilent microarray data}
\usage{
gplotKMeans(data, centers = 2, methodDist = "euclidean", groups = NULL,
  getresult = FALSE, old = FALSE, ...)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{centers}{either the number of clusters, say \emph{k}, or a set initial (distinct)
cluster centers. If a number, a random set of (distinct) rows in x is chosen as initial centres.
See \code{\link{kmeans}}. Default: \code{2}.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{groups}{forms by groups. Only if argument \code{old} is \code{FALSE}.}

\item{getresult}{print result K-Means clustering}

\item{old}{if TRUE draw by plot, else ggplot. Default:{FALSE}}
}
\description{
Clustering MDS by K-Means clustering algorithm.
}

