% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotMA.R
\name{gplotMA}
\alias{gplotMA}
\title{MA-plot: plot differences versus averages for Agilent microarray data}
\usage{
gplotMA(data, array = 1, columnProbe = "ControlType", probe = NULL,
  alpha = 0.5, size = 2)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{array}{number array or name array to be displayed. Default \code{1}.}

\item{columnProbe}{column where contain probe number}

\item{probe}{agilent type of probe. See in "ControlType" column. Default \code{NULL}, i.e. all.}

\item{alpha}{transparent points (min = 1, max = 0). Default \code{0.25}.}

\item{size}{size of points. Default \code{2}.}
}
\value{
A plot is created on the current graphics device.
}
\description{
A generic function which produces an MA-plot for an object containing microarray data.
}
\details{
An MA-plot is a plot of log-intensity ratios (M-values) versus log-intensity averages 
(A-values). If MA is an RGList or MAList then this function produces an ordinary within-array 
MA-plot. If MA is an MArrayLM object, then the plot is an fitted model MA-plot in which the 
estimated coefficient is on the y-axis and the average A-value is on the x-axis.
}

