% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotOutliners.R
\name{gplotOutliners}
\alias{gplotOutliners}
\title{Plot outliners in array}
\usage{
gplotOutliners(data, smp = 1, flags = TRUE)
}
\arguments{
\item{data}{RGList, EListRaw}

\item{smp}{array}

\item{flags}{logical, if \code{TRUE} find outliners by flags else by ratio of mean and median}
}
\description{
Plot outliners in array
}

