% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplotRC.R
\name{gplotRC}
\alias{gplotRC}
\title{Plot Row&Column}
\usage{
gplotRC(data, smp = 1)
}
\arguments{
\item{data}{RGarray, Earray}

\item{smp}{array}
}
\description{
Plot of median distribution by rows and columns
}

