% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gplotSNP.R
\name{gplotSNP}
\alias{gplotSNP}
\title{Plot SNP distribution}
\usage{
gplotSNP(data, array = 1, reference = c("none", "male", "female"),
  type = c("log", "theta"), SNPprobe = -15000)
}
\arguments{
\item{data}{MAList or RGList with SNP probes}

\item{array}{number of array}

\item{reference}{matrix, data of cut/uncut alleles}

\item{type}{type of SNP ditribution: \code{log} or \code{theta}}

\item{SNPprobe}{numeric, type number of SNP probes. Default: -15000}
}
\description{
Plot SNP distribution
}
\details{
Argument \code{array} might be numeric, sequence number of array or
parametre \code{"all"} to plot all samples in one plot.
}

