% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limarg.R
\name{limarg}
\alias{limarg}
\title{Limarg}
\usage{
limarg(arg = NULL, maxi = NULL, maxl = NULL, class = c("numeric",
  "integer", "double"))
}
\arguments{
\item{arg}{numeric or vector, argument}

\item{maxi}{check argument with maximums}

\item{maxl}{maxwell}

\item{class}{check classes}
}
\description{
Internal function. Limits of arguments
}

