% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSmall.R
\name{makeSmall}
\alias{makeSmall}
\title{Make small array}
\usage{
makeSmall(data)
}
\arguments{
\item{data}{matrix}
}
\value{
The size of reduced matrix is determined by the Greatest Common Divisor. 
Reduced values will be aligned by median.
}
\description{
The reduction is based on the Greatest Common Divisor and median
}
\keyword{internal}

