% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseReductionCGH.R
\name{noiseReductionCGH}
\alias{noiseReductionCGH}
\title{Noise reduction for aCGH}
\usage{
noiseReductionCGH(data, N = NULL, method = c("mad", "iqr"))
}
\arguments{
\item{data}{RGList only}

\item{N}{numeric, i}

\item{method}{iqr or mad}
}
\description{
Noise reduction only for array CGH data
}

