% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDG.R
\name{plotDG}
\alias{plotDG}
\title{Plot dendrogramm for Agilent microarray data}
\usage{
plotDG(data, methodDist = "euclidean", methodHclust = "complete", ...)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{methodHclust}{method for hierarchical cluster analysis. See \code{\link{hclust}}. Default: {"complete"}.}
}
\description{
Clustering tree by Hierarchical clustering algorithm.
}

