% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Plot clustering MDS for Agilent microarray data}
\usage{
plotKM(data, centers = 2, methodDist = "euclidean", print = FALSE)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{centers}{either the number of clusters, say \emph{k}, or a set initial (distinct)
cluster centers. If a number, a random set of (distinct) rows in x is chosen as initial centres.
See \code{\link{kmeans}}. Default: \code{2}.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{print}{print result K-Means clustering}
}
\description{
Clustering MDS by K-Means clustering algorithm.
}

