% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readAgilent.R
\name{readAgilent}
\alias{readAgilent}
\title{Read agilent files}
\usage{
readAgilent(targets = NULL, path = getwd(), format = c("compact", "full"),
  macolors = 2, collect = c("processed", "median", "mean", "LogRatio",
  "DyeNorm"))
}
\arguments{
\item{targets}{Feature Extraction's files. TXT format files only.}

\item{path}{filepath, where FE-files are stored. Default: getwd().}

\item{format}{which format your FE-files have? Default: \code{"compact"}}

\item{macolors}{read two-color (2) or one-color (1) microarray.}

\item{collect}{what the data signals you want to process? Default: \code{"processed"}.}
}
\value{
Read Feature Extraction's files.
}
\description{
Read Feature Extraction's txt files.
}
\details{
Collecting data from the format \code{"compact"} can be taken on the processed signals
(\code{"processed"}), raw data (\code{"mean"} or \code{"median"}) or LogRatio (\code{"LogRatio"}).
At the same time, if format selected as \code{"full"} you may collect normalized data (\code{"DyeNorm"}).
}
\examples{
# targets <- read.table("targets.txt", header = TRUE, stringsAsFactors = FALSE)
# MyArrays <- readAgilent(targets, collect = "median")
}
\seealso{
\code{\link{readGenePix}}
}

