% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGenePix.R
\name{readGenePix}
\alias{readGenePix}
\title{Read genepix files}
\usage{
readGenePix(targets = NULL, path = getwd(), custom.data = NULL,
  custom.ann = NULL)
}
\arguments{
\item{targets}{GenePix's files. GPR format files only.}

\item{path}{filepath, where GP-files are stored. Default: getwd().}

\item{custom.data}{custom data. For example \code{c(E = "F532 Median", Eb = "F532 MedianBG")}}

\item{custom.ann}{custom annotation. For example \code{c(Row="Row", Col="Column")}}
}
\description{
Read GenePix grp files.
}
\details{
Targets object must be class \code{data.frame}, and have two required column \code{name} 
(eg.: "array1", "array2") and \code{filenames} (eg. "2802860_12158125.gpr").
}
\examples{
# targets <- read.table("targets.txt", header = TRUE, stringsAsFactors = FALSE)
# MyArrays <- readGenePix(targets)

}
\seealso{
\code{\link{readAgilent}}
}

