% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmOutliners.R
\name{rmOutliners}
\alias{rmOutliners}
\title{Remove Outliners}
\usage{
rmOutliners(data, flags = TRUE, method = c("intra", "between"))
}
\arguments{
\item{data}{RGList, EListRaw}

\item{flags}{logical, if \code{TRUE} find outliners by flags else by ratio of mean and median}

\item{method}{if \code{intra} - detect and remove outliners intra-array, if \code{between} -
detect and remove outliners inter-array.}
}
\description{
Detect and remove outliners by intra-array or inter-array methods.
}

