% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmByCluster.R
\name{segmByCluster}
\alias{segmByCluster}
\title{Multicore Genome Segmentation Program}
\usage{
segmByCluster(x, weights = NULL, alpha = 0.01, nperm = 10000,
  p.method = c("hybrid", "perm"), min.width = 2, kmax = 25, nmin = 200,
  eta = 0.05, sbdry = NULL, trim = 0.025, undo.splits = c("none",
  "prune", "sdundo"), undo.prune = 0.05, undo.SD = 2, verbose = 1,
  cluster = getDoParWorkers())
}
\arguments{
\item{x}{an object of class CNA}

\item{weights}{a vector of weights for the probes. The weights should be inversely proportional to their variances. Currently all weights should be positive i.e. remove probes with zero weight prior to segmentation.}

\item{alpha}{significance levels for the test to accept change-points.}

\item{nperm}{number of permutations used for p-value computation.}

\item{p.method}{method used for p-value computation. For the "perm" method the p-value is based on full permutation. For the "hybrid" method the maximum over the entire region is split into maximum of max over small segments and max over the rest. Approximation is used for the larger segment max. Default is hybrid.}

\item{min.width}{the minimum number of markers for a changed segment. The default is 2 but can be made larger. Maximum possible value is set at 5 since arbitrary widths can have the undesirable effect of incorrect change-points when a true signal of narrow widths exists.}

\item{kmax}{the maximum width of smaller segment for permutation in the hybrid method.}

\item{nmin}{the minimum length of data for which the approximation of maximum statistic is used under the hybrid method. should be larger than 4*kmax}

\item{eta}{the probability to declare a change conditioned on the permuted statistic exceeding the observed statistic exactly j (= 1,...,nperm*alpha) times.}

\item{sbdry}{the sequential boundary used to stop and declare a change. This boundary is a function of nperm, alpha and eta. It can be obtained using the function "getbdry" and used instead of having the "segment" function compute it every time it is called.}

\item{trim}{proportion of data to be trimmed for variance calculation for smoothing outliers and undoing splits based on SD.}

\item{undo.splits}{A character string specifying how change-points are to be undone, if at all. Default is "none". Other choices are "prune", which uses a sum of squares criterion, and "sdundo", which undoes splits that are not at least this many SDs apart.}

\item{undo.prune}{the proportional increase in sum of squares allowed when eliminating splits if undo.splits="prune".}

\item{undo.SD}{the number of SDs between means to keep a split if undo.splits="sdundo".}

\item{verbose}{level of verbosity for monitoring the program's progress where 0 produces no printout, 1 prints the current sample, 2 the current chromosome and 3 the current segment. The default level is 1.}

\item{cluster}{A specification appropriate to the type of cluster.}
}
\value{
See details \code{\link{segment}} and \code{\link{makeCluster}}
}
\description{
This program segments DNA copy number 
data into regions of estimated 
equal copy number using circular binary segmentation (CBS) (\code{\link{DNAcopy}} package)
and multicore (\code{\link{doParallel}}).
}

