% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrCalc.R
\name{thrCalc}
\alias{thrCalc}
\title{Calculate threshold}
\usage{
thrCalc(data, method = c("MAD", "SD"), n = 1)
}
\arguments{
\item{data}{subset of CBS}

\item{method}{perform by \code{MAD} or \code{SD}}

\item{numeric, }{multiple to method}
}
\description{
find threshold
}

