% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDG.R
\name{plotDG}
\alias{plotDG}
\title{Plot dendrogramm for Agilent microarray data}
\usage{
plotDG(data, cyanine = c("R", "G"), centers = 2, methodDist = "euclidean",
  methodHclust = "complete", print = FALSE)

plotDG(data, cyanine = c("R", "G"), centers = 2, methodDist = "euclidean",
  methodHclust = "complete", print = FALSE)
}
\arguments{
\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{cyanine}{only for RGList data class.}

\item{centers}{either the number of clusters, say \emph{k}, or a set initial (distinct)
cluster centers. If a number, a random set of (distinct) rows in x is chosen as initial centres.
See \code{\link{kmeans}}. Default: \code{2}.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{methodHclust}{method for hierarchical cluster analysis. See \code{\link{hclust}}. Default: {"complete"}.}

\item{print}{print result K-Means clustering}

\item{data}{microarray data. Only for \code{RGlist}, \code{MAlist} or \code{ELits} class objects.}

\item{cyanine}{only for RGList data class.}

\item{methodDist}{method for Distance Matrix Computation. See \code{\link{dist}}. Default: {"euclidean"}.}

\item{methodHclust}{method for hierarchical cluster analysis. See \code{\link{hclust}}. Default: {"complete"}.}
}
\description{
Clustering tree by Hierarchical clustering algorithm.

Clustering MDS by K-Means clustering algorithm.
}

