% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAgilent.R
\name{readAgilent}
\alias{readAgilent}
\title{Read agilent files}
\usage{
readAgilent(targets = NULL, path = getwd(), format = c("C", "F"),
  dyes = 2, collect = c("proc", "median", "mean", "lr", "dn"),
  verbose = 0)
}
\arguments{
\item{targets}{Feature Extraction's files. TXT format files only.}

\item{path}{filepath, where FE-files are stored. Default: getwd().}

\item{format}{which format your FE-files have? Default: \code{"compact"}}

\item{collect}{what the data signals you want to process? Default: \code{"processed"}.}

\item{macolors}{read two-color (2) or one-color (1) microarray.}
}
\value{
Read Feature Extraction's files.
}
\description{
Read Feature Extraction's txt files.
}
\details{
Collecting data from the format \code{"compact"} can be taken on the processed signals
(\code{"processed"}), raw data (\code{"mean"} or \code{"median"}) or LogRatio (\code{"LogRatio"}).
At the same time, if format selected as \code{"full"} you may collect normalized data (\code{"DyeNorm"}).

Targets object must be class \code{data.frame}, and have two required column \code{name} 
(eg.: "array1", "array2") and \code{filenames} (eg. "2802860_12158125.txt").
}
\examples{
# targets <- read.table("targets.txt", header = TRUE, stringsAsFactors = FALSE)
# MyArrays <- readAgilent(targets, collect = "median")

}
\seealso{
\code{\link{readGenePix}}
}

